罗永明整合代码0.5

+++ boshen/src/main/java/com/gpnu/boshen/controller/ConsultCategoryController.java
@@ -1,7 +1,10 @@
 package com.gpnu.boshen.controller;
 
 import com.gpnu.boshen.entity.ConsultCategory;
+import com.gpnu.boshen.enums.CommentStateEnum;
+import com.gpnu.boshen.enums.ConsultCategoryStateEnum;
 import com.gpnu.boshen.mapper.ConsultCategoryMapper;
+import com.gpnu.boshen.vo.ResultVo;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
@@ -9,7 +12,7 @@ import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
 
-@Controller
+@RestController
 public class ConsultCategoryController {
 
     @Autowired
@@ -20,8 +23,8 @@ public class ConsultCategoryController {
      * @return
      */
     @GetMapping("/consult/categories")
-    public List<ConsultCategory> list() {
-        return consultCategoryMapper.list();
+    public ResultVo list() {
+        return new ResultVo(ConsultCategoryStateEnum.SUCCESS,consultCategoryMapper.list()) ;
     }
 
     /**
@@ -30,22 +33,38 @@ public class ConsultCategoryController {
      * @return
      */
     @PostMapping("/consult/category")
-    public ConsultCategory insert(ConsultCategory category) {
-        return consultCategoryMapper.insert(category);
+    public ResultVo insert(@RequestBody ConsultCategory category) {
+        if (category.getConsultCategoryName() == null){
+            return new ResultVo(ConsultCategoryStateEnum.FAIL_NULL_PARAM);
+        }
+        consultCategoryMapper.insert(category);
+        return new ResultVo(ConsultCategoryStateEnum.SUCCESS);
     }
 
     /**
      * 根据id 查询咨询类别
      * @param id
-     * @param model
      * @return
      */
     @GetMapping("/consult/category/{id}")
-    public String toEditPage(@PathVariable("id") int id, Model model) {
+    public ResultVo findById(@PathVariable("id") Integer id) {
+        if (id == null) {
+            return new ResultVo(CommentStateEnum.FAIL_NULL_PARAM);
+        }
         ConsultCategory consultCategory = consultCategoryMapper.get(id);
-        model.addAttribute("consultCategory", consultCategory);
+        return new ResultVo(ConsultCategoryStateEnum.SUCCESS,consultCategory);
+    }
 
-        return "/consult/category/add";
+    /**
+     * 改询咨询类别
+     */
+    @PutMapping("/consult/category")
+    public ResultVo update(@RequestBody ConsultCategory category){
+        if (category.getConsultCategoryName() == null || category.getConsultCategoryId() == null){
+            return new ResultVo(CommentStateEnum.FAIL_NULL_PARAM);
+        }
+        consultCategoryMapper.update(category);
+        return new ResultVo(ConsultCategoryStateEnum.SUCCESS);
     }
 
     /**
@@ -54,7 +73,11 @@ public class ConsultCategoryController {
      * @return
      */
     @DeleteMapping("/consult/category/{id}")
-    public ConsultCategory delete(@PathVariable("id") int id) {
-        return consultCategoryMapper.delete(id);
+    public ResultVo delete(@PathVariable("id") Integer id) {
+        if (id == null){
+            return new ResultVo(CommentStateEnum.FAIL_NULL_PARAM);
+        }
+        consultCategoryMapper.delete(id);
+        return new ResultVo(ConsultCategoryStateEnum.SUCCESS);
     }
 }
